# docs/api-audit/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tecbunny API Surface",
    "version": "1.0.0",
    "description": "Generated from repository source by scripts/audit-api-surface.mjs."
  },
  "servers": [
    {
      "url": "{{API_BASE_URL}}"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/superadmin/areas": {
      "delete": {
        "description": "DELETE /api/superadmin/areas",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/superadmin/src/app/api/superadmin/areas/route.ts",
        "x-database-tables": [
          "area_pincodes",
          "areas",
          "profiles",
          "user_area_assignments"
        ],
        "x-frontend-usage": [
          "apps/superadmin/src/app/superadmin/mgmt/areas/page.tsx"
        ]
      },
      "get": {
        "description": "GET /api/superadmin/areas",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/superadmin/src/app/api/superadmin/areas/route.ts",
        "x-database-tables": [
          "area_pincodes",
          "areas",
          "profiles",
          "user_area_assignments"
        ],
        "x-frontend-usage": []
      },
      "post": {
        "description": "POST /api/superadmin/areas",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/superadmin/src/app/api/superadmin/areas/route.ts",
        "x-database-tables": [
          "area_pincodes",
          "areas",
          "profiles",
          "user_area_assignments"
        ],
        "x-frontend-usage": []
      }
    },
    "/api/superadmin/catalogue/generate": {
      "post": {
        "description": "POST /api/superadmin/catalogue/generate",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Observed or expected 200 response"
          },
          "400": {
            "description": "Observed or expected 400 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/superadmin/src/app/api/superadmin/catalogue/generate/route.ts",
        "x-database-tables": [
          "products",
          "services"
        ],
        "x-frontend-usage": [
          "apps/superadmin/src/app/superadmin/mgmt/catalogue/page.tsx"
        ]
      }
    },
    "/api/superadmin/custom-setup-offers": {
      "delete": {
        "description": "DELETE /api/superadmin/custom-setup-offers",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [],
        "x-route-file": "apps/superadmin/src/app/api/superadmin/custom-setup-offers/route.ts",
        "x-database-tables": [
          "custom_setup_offers"
        ],
        "x-frontend-usage": [
          "apps/superadmin/src/components/superadmin/CustomSetupOffersManager.tsx"
        ]
      },
      "get": {
        "description": "GET /api/superadmin/custom-setup-offers",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [],
        "x-route-file": "apps/superadmin/src/app/api/superadmin/custom-setup-offers/route.ts",
        "x-database-tables": [
          "custom_setup_offers"
        ],
        "x-frontend-usage": []
      },
      "post": {
        "description": "POST /api/superadmin/custom-setup-offers",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Observed or expected 201 response"
          },
          "400": {
            "description": "Observed or expected 400 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [],
        "x-route-file": "apps/superadmin/src/app/api/superadmin/custom-setup-offers/route.ts",
        "x-database-tables": [
          "custom_setup_offers"
        ],
        "x-frontend-usage": []
      },
      "put": {
        "description": "PUT /api/superadmin/custom-setup-offers",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [],
        "x-route-file": "apps/superadmin/src/app/api/superadmin/custom-setup-offers/route.ts",
        "x-database-tables": [
          "custom_setup_offers"
        ],
        "x-frontend-usage": []
      }
    },
...
  }
}